Inline remaining one use of getViewConfigured
