+++ packages/dds/tree/src/test/shared-tree/treeCheckout.spec.ts
@@ -19,8 +19,13 @@ import {
 	EmptyKey,
 	type RevertibleFactory,
 	type NormalizedFieldUpPath,
+	TreeStoredSchemaRepository,
 } from "../../core/index.js";
-import { FieldKinds } from "../../feature-libraries/index.js";
+import {
+	buildForest,
+	FieldKinds,
+	MockNodeIdentifierManager,
+} from "../../feature-libraries/index.js";
 import {
 	getBranch,
 	Tree,
@@ -28,6 +33,7 @@ import {
 	type ITreeCheckout,
 	type ITreeCheckoutFork,
 	type BranchableTree,
+	createTreeCheckout,
 } from "../../shared-tree/index.js";
 import {
 	TestTreeProviderLite,
@@ -35,7 +41,9 @@ import {
 	createTestUndoRedoStacks,
 	expectSchemaEqual,
 	getView,
-	getViewConfigured,
+	mintRevisionTag,
+	testIdCompressor,
+	testRevisionTagCodec,
 	validateUsageError,
 	viewCheckout,
 } from "../utils.js";
@@ -1451,7 +1459,23 @@ function itView<
 		logger: IMockLoggerExt;
 	} {
 		const logger = createMockLoggerExt();
-		const view = getViewConfigured(config, undefined, logger);
+
+		const checkout = createTreeCheckout(
+			testIdCompressor,
+			mintRevisionTag,
+			testRevisionTagCodec,
+			{
+				forest: buildForest(),
+				schema: new TreeStoredSchemaRepository(),
+				logger,
+			},
+		);
+		const view = new SchematizingSimpleTreeView<TRootSchema>(
+			checkout,
+			config,
+			new MockNodeIdentifierManager(),
+		);
+
 		if (fork) {
 			const treeBranch = getBranch(view).branch();
 			const viewBranch = treeBranch.viewWith(view.config);
